test(jangar): align AgentRun goal hash coverage
